+++ fxn/__init__.py
@@ -1,8 +1,9 @@
 # 
 #   Function
-#   Copyright © 2024 NatML Inc. All Rights Reserved.
+#   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
+from .compile import *
 from .function import Function
 from .types import *
 from .version import *