+++ fxn/beta/__init__.py
@@ -2,3 +2,5 @@
 #   Function
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
+
+from .remote import RemoteAcceleration