+++ fxn/types/__init__.py
@@ -1,9 +1,9 @@
 # 
 #   Function
-#   Copyright © 2024 NatML Inc. All Rights Reserved.
+#   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
 from .dtype import Dtype
-from .prediction import Prediction, PredictionResource
-from .predictor import Acceleration, AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, Signature
+from .prediction import Acceleration, Prediction, PredictionResource
+from .predictor import AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, Signature
 from .user import User