+++ fxn/types/prediction.py
@@ -1,8 +1,9 @@
 #
 #   Function
-#   Copyright © 2024 NatML Inc. All Rights Reserved.
+#   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
+from enum import IntFlag
 from pydantic import BaseModel, Field
 from typing import Any
 
@@ -43,3 +44,12 @@ class Prediction (BaseModel):
     error: str | None = Field(default=None, description="Prediction error. This is `None` if the prediction completed successfully.")
     logs: str | None = Field(default=None, description="Prediction logs.")
     created: str = Field(description="Date created.")
+
+class Acceleration (IntFlag):
+    """
+    Predictor acceleration.
+    """
+    Auto    = 0,
+    CPU     = 1 << 0,
+    GPU     = 1 << 1,
+    NPU     = 1 << 2