+++ fxnc.py
@@ -1,6 +1,6 @@
 # 
 #   Function
-#   Copyright © 2024 NatML Inc. All Rights Reserved.
+#   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
 from argparse import ArgumentParser